Clear password field in iframe when popup opens
e6c1c19a73c98a6d334f31772614c47da0825066 commit did not clear the password field if it was in an iframe.  The easiest way to make it work is to send the message to all frames.  That means I'm sending the userid to frames that might not be trusted.  I don't think that's a big deal because the message goes only to my content script.