Update instructions for shared credentials
Groups of sites in sharedCredentials.js share the same userid and password, e.g., hulu.com, disney.com, and espn.com.  That means they must use the same nickname, which could be confusing for users.

This commit includes instructions explaining that you setup any one of the sites in a set, and the others in that set all get the same settings without any user interaction.