Solve an update problem
I was updating certain values when the relevant field loses focus.  However, the user can mouse out of the popup without ever changing the focus of the field being edited, resulting in the changes being lost.

This commit guarantees the updates will be made by forcing a focus on the main panel, which triggers a blur event on the field being edited.  See the comment starting on line 399 for a potential hazard.