Use const instead of let for bookmark folder
I was using let to assign a value to SitePasswordData because I want to be able to reset it for test, debug, or demo use.  However, once it's set, I don't want it changed due to a bug.  I can't use Object.freeze because it only works for objects and arrays, but the value is a string.