Fix dblclick for user name
when there is more than one input element on the page.  I wasn't registering the dblclick handler unless there was exactly 1 text field on the page in order to avoid confusion.  However, the user is unlikely to try to insert a user name into an irrelevant input element, such as a search field.  There's little risk even if that's done.