Don't set placeholder for dblclick username
It looks odd when the placeholder for a random text input field, such as a search input, says to dblclick for your username.  I'm still leaving the tooltip, which I hope won't be too confusing.

I also updated the instructions to explain that dblclick inserts your user name on most text input fields.