Fix input field problem
Say that there's a password field with two text input fields above it.  SitePassword assumes the one closest to the password field is for the username.  When that's not the case, the user needs to type over what's there.  The problem was that SitePassword kept putting the username back on every keystroke.

The solution is to change the field only if it's empty.  Note that both tests are needed.  The fillfield() test doesn't work if the user cleared the field.  The mutationObserver test doesn't work if the call comes from somewhere else.

I'm not entirely happy with this approach, so the next commit may do something different.