Prevent a clickjacking attack
I don't put the click handler on a password field unless it's visible.  One of the visibility checks is opacity, since there's no good reason to make it <1 on a password field.  The tricky thing is that an element inherits the visibility of its ancestors.  I thought that's wha getComputedStyle() did, but it turns out that I have to check each ancestor individually.