Defeat clickjacking attack
There are certain DOM elements, namely popover and dialog, that appear in a special top layer that obscures elements under them.  Unfortunately, clicks to those invisible elements reach them, which can be used in a clickjacking attack.

This commit treats the username and password fields found by findpw.js as hidden when that happens.  If that heuristic is wrong, the only downside is that the user will have to manually enter or paste into those fields.