Change userid heuristic
I resisted relying on the content of the id, autocomplete, and label components because of pages in other languages.  However, the Lowes create account page forced me to cave.  It's got a phone number input just before the password field.  Trying to fill in that field with a username confuses the Lowes JavaScript that verifies that it's a phone number.

I'm now using a Copilot generated function that implements a heuristic that relies on English in the HTML.  I added one test of my own; if there's only 1 visible text input element on the page, assume it takes a username.