Help with debugging content script
When single stepping the debugger on a content script, the page loses focus.  However, I'm using page focus to decide if the page should send messages.  So, if I'm single stepping the content script, it won't send messages.  The problem is I need to see the response.

My solution is to use debugMode to turn off that behavior.  Now the problem is that all pages can send messages even without the focus.  The result is interleaved messages in bg.js that mess stuff up.  My solution is to use debug mode to turn off automatic updating when the content script changes.  I can still see the changes by reloading the page I'm working on.

I don't like that I have to set debug mode separately in bg.js and findpw.js, so I might change how I do that later.