Fix fill fields test
This fix goes back to an error in commit dcf7b83.  I was trying to fix a problem at truecar.com where your username was showing up in a search field on a page with a single password element.  However, I applied that test to password fields, too, which meant the password didn't get filled in.