Fix visibility color test
A border may not have an assigned color, but it still contrasts with the field and background.  I'm going to assume that the border is a different color than the field color, which means either the border or the field will show up against the background.

Notice the test for a border.  It's possible to set all kinds of border properties without actually setting style.border.