Got rid of a race condition
by eliminating a global variable.  cpi, the result of calling countpwid() was a global variable so I didn't have to redo the expensive search for a password field so often.  Unfortunately, countpwid() could be called while another event was using the result of a previous call.  The result was unexpected failures depending on the timing of events.

It turns out that most of the time the function is called from a mutation observer, which does need to redo the calculation.  My solution then is to make cpi a local variable and call countpwid() every time cpi is needed.  The result is no more confusion (at least from this cause).

I also tried using a promise queue to serialize calls to countpwid(), but it was more complicated because it's called from several event handlers.