Avoid a race
It doesn't happen often, but there are rare times when a page mutation causes the value of cpi to change while it's being used.  The solution is to make cpi a local variable.  The problem is extensionRemoved(), which is called by countpw() while calling countpw(), resulting in a stack overflow.  The solution is to introduce a global lastcpi that is only used by extensionRemoved().