Fix handling of mutations
I wasn't handling mutations properly.  As a result, I was in danger of mutation cycling, where my changes triggered a mutation that resulted in more of my changes.  The fix allowed me to reduce the number of times I call countpwid().

This version only calls countpwid() if the content script is running on the active tab, which it learns from the background script when it transitions to being active.  As a result, the content script consumes very few resources when it's not on the active tab.