Toggle view password
Some sites have an icon for seeing your password, but many don't.  I want to be able to show the password on all sites for the demo, but it might be useful in general.

Some sites don't let me toggle back to type password,  e.g., (usps.com), supposedly for security purposes.