Put back code I deleted for some reason
There's a heuristic for finding the likely username field.  You start at the password field (if there's only 1), and walk up the DOM to find the first visible text input field.  I had deleted that test (for a reason I can't find) that filled in the username in a text field nowhere near a password field at usps.com.