Fix isUsernameField
Most web pages put an id or a name on each input element, but some don't.  In that case, I just have to assume that the text input element could hold a username.  In other words, instead of using the attributes to say the element takes a username, I use the attributes to say that it isn't.