Fix how I find the username field
I was testing a text input field against a set of attributes to see if it holds a username.  Unfortunately, not all sites use an attribute in my list, so I didn't fill in the username even on a very simple site with one username and one password field.

I'm going back to looking for the first visible text input field before the password field in the DOM.