Add ability to reload all tabs in demo window
I need to reload all the tabs in the demo window as part of the setup.  I can now do that with CTRL-r in any tab in that window, but only if demoMode === true in bg.js.  This way users won't be confused if CTRL-r is used by some random web page.