Fix inconsistency in download site data
Downloading site data produces an HTML file with links you can click to take you to sitepassword.info with the form filled in with the site's settings.  Unfortunately, the domain name that appears on the web site wasn't necessarily the one you clicked if there are more than one domain name for that account.  This commit fixes that inconsistency.