Correctly handle login form in an iframe
When the login form is in an iframe, I was returning the password for blank site and user names.  Unfortunately, the user didn't know that if the password was shown as dots.

Now I return a blank site password, which tells the content script to show an alert telling the user to click the SitePassword icon when the user clicks on the password field.  The only minor glitch is that the tooltip says to Click Here.