Properly handle login form in an iframe
I wasn't using the domain name of the iframe when calculating the site password.  That's bad.  Imagine an ad with a login form on your bank's login page.  (Not good practice by them, but you never know.)  The ad would get your bank password.  Oops.

I struggled mightily to come up with a fix. Some thinks I tried were very complicated, others didn't work at all.

As you can see from this commit, the changes are modest.  The secret is returning "" when the content script sends a getPassword message when the form is filled in, but there is no metadata for the iframe domain name.  I use that to tell the user to open the popup, which displays the phishing warning.