Mostly get rid of bg.settings.domainname
The domain name and pw domain name do not belong in the settings, since they're different for each domain that has a login form for the account.  I gotten rid of most of those uses, but there's one in each of bg.js and ssp.js that doesn't work when the login form is in an iframe.

I'm making this commit so it's easier for me to isolate the problem.