Handle popup-like windows
Popup-like windows do not show extension icons, so there's no way to open the popup.  They are also used for browser-in-the-browser phishing attacks, so SitePassword can't give a phishing warning.

My solution is to open a alert telling the user to open that URL in a regular tab.  Everything works from there.  I also added a couple of paragraphs in the document to explain that.